Function naive_parsec::combinators::map_result

pub fn map_result<'f, A: 'f, B: 'f, E: ToString, S: Clone>(
    p: impl Parsable<Stream = S, Result = A> + 'f, 
    f: impl Fn(A) -> Result<B, E> + 'f
) -> Parser<'f, B, S>

Combinator: map_result (function ver.)

Maps the result of a parser to another value using the given function that produces an Result<T, E>. The only difference with map is that map_result will automatically try to unwrap the Result<T, E>. If an Err is yeilded, map_result will log down the error message. Therefore, it requires E from Result<T, E> to implement ToString trait.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{CharStream, satisfy};

// A parser that consumes a natural number
let parser = map_result(
                some(satisfy(|&ch| ch.is_digit(10))),
                |v| v.into_iter().collect::<String>().parse::<i64>()
             );

let mut st = CharStream::new("12345");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(12345), res);
assert_eq!("", st.as_str());
assert_eq!(0, logs.len());