Function naive_parsec::combinators::right

pub fn right<'f, A: 'f, B: 'f, S: Clone + 'f>(
    p1: impl Parsable<Stream = S, Result = A> + 'f, 
    p2: impl Parsable<Stream = S, Result = B> + 'f
) -> Parser<'f, B, S>

Combinator: right (function ver.)

A sequential combinator that applys the first parser then the second. If both parsers succeed, returns the result of the second parser, otherwise fail.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Consume a character 'A', then a character 'B'
let parser = right(char('A'), char('B'));

let mut st = CharStream::new("ABC");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('B'), res);
assert_eq!("C", st.as_str());
assert_eq!(0, logs.len());