Trait naive_parsec::combinators::BindExt

pub trait BindExt<'f, A: 'f, S>: Parsable<Stream = S, Result = A> {
    fn bind<B: 'f, P>(self, f: impl Fn(A) -> P + 'f) -> Parser<'f, B, S>
    where
        P: Parsable<Stream = S, Result = B>,
        S: Clone,
        Self: Sized + 'f,
    { ... }
}

Implement bind combinator for Parsable<S>.

Provided methods

fn bind<B: 'f, P>(self, f: impl Fn(A) -> P + 'f) -> Parser<'f, B, S> where
P: Parsable<Stream = S, Result = B>,
S: Clone,
Self: Sized + 'f,

Combinator: bind

Monadic bind operator (>>=) for context sensitive parsing.

Properties

Should satisfy Monad laws:

Left-identity: pure(x).bind(f) ~ f(x)

Right-identity: p.bind(|x| pure(x)) ~ p

Associativity: p.bind(f).bind(g) ~ p.bind(|x| f(x).bind(g))

Check out test_bind module in the source code for naive examples of above laws.

Example

The code example below parses expr with the following grammar:

use naive_parsec::core::Parsable;
use naive_parsec::combinators::BindExt;
use naive_parsec::primitives::{ CharStream, char, satisfy };

// <expr> := <uppercase_letter> '+'
// <expr> := <lowercase_letter> '-'
let parser = satisfy(|_| true)
           .bind(|ch| if ch.is_uppercase() {
               char('+')
           } else {
               char('-')
           });

let (res1, _) = parser.exec(&mut CharStream::new("A+"));
assert_eq!(Some('+'), res1);
let (res2, _) = parser.exec(&mut CharStream::new("a-"));
assert_eq!(Some('-'), res2);

Loading content...