Trait naive_parsec::combinators::OrExt

pub trait OrExt<'f, A: 'f, S>: Parsable<Stream = S, Result = A> {
    fn or(
        self, 
        p: impl Parsable<Stream = S, Result = A> + 'f
    ) -> Parser<'f, A, S>
    where
        S: Clone,
        Self: Sized + 'f,
    { ... }
}

Implement or combinator for Parsable<S>.

Provided methods

fn or(self, p: impl Parsable<Stream = S, Result = A> + 'f) -> Parser<'f, A, S> where
S: Clone,
Self: Sized + 'f,

Combinator: or (function ver.)

Alternative combinator. Accepts two parsers as arguments, if the first parser succeeds then its result is returned, otherwise the result of the second parser is returned.

Examples

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Comsumes a character 'A' or a character 'B'
let parser = char('B').or(char('A'));

let mut st = CharStream::new("Ahhh");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('A'), res);
assert_eq!("hhh", st.as_str());
assert_eq!(0, logs.len());

Loading content...