Trait naive_parsec::combinators::SequentialExt

pub trait SequentialExt<'f, A: 'f, S>: Parsable<Stream = S, Result = A> {
    fn and<B: 'f>(
        self, 
        p: impl Parsable<Stream = S, Result = B> + 'f
    ) -> Parser<'f, (A, B), S>
    where
        Self: Sized + 'f,
        S: Clone,
    { ... }

    fn left<B: 'f>(
        self, 
        p: impl Parsable<Stream = S, Result = B> + 'f
    ) -> Parser<'f, A, S>
    where
        Self: Sized + 'f,
        S: Clone + 'f,
    { ... }

    fn right<B: 'f>(
        self, 
        p: impl Parsable<Stream = S, Result = B> + 'f
    ) -> Parser<'f, B, S>
    where
        Self: Sized + 'f,
        S: Clone + 'f,
    { ... }

    fn mid<B: 'f, C: 'f>(
        self, 
        p1: impl Parsable<Stream = S, Result = B> + 'f, 
        p2: impl Parsable<Stream = S, Result = C> + 'f
    ) -> Parser<'f, B, S>
    where
        Self: Sized + 'f,
        S: Clone + 'f,
    { ... }
}

Implement sequential combinators for Parsable<S>.

Provided methods

fn and<B: 'f>(
self,
p: impl Parsable<Stream = S, Result = B> + 'f
) -> Parser<'f, (A, B), S> where
Self: Sized + 'f,
S: Clone,

Combinator: and (function ver.)

A sequential combinator that applys the first parser then the second. If both parsers succeed, returns a tuple containing their results, otherwise fail.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Consume a character 'A', then a character 'B'
let parser = char('A').and(char('B'));

let mut st = CharStream::new("ABC");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(('A', 'B')), res);
assert_eq!("C", st.as_str());
assert_eq!(0, logs.len());

fn left<B: 'f>(
self,
p: impl Parsable<Stream = S, Result = B> + 'f
) -> Parser<'f, A, S> where
Self: Sized + 'f,
S: Clone + 'f,

Combinator: left

A sequential combinator that applys the first parser then the second. If both parsers succeed, returns the result of the first parser, otherwise fail.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Consume a character 'A', then a character 'B'
let parser = char('A').left(char('B'));

let mut st = CharStream::new("ABC");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('A'), res);
assert_eq!("C", st.as_str());
assert_eq!(0, logs.len());

fn right<B: 'f>(
self,
p: impl Parsable<Stream = S, Result = B> + 'f
) -> Parser<'f, B, S> where
Self: Sized + 'f,
S: Clone + 'f,

Combinator: right

A sequential combinator that applys the first parser then the second. If both parsers succeed, returns the result of the second parser, otherwise fail.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Consume a character 'A', then a character 'B'
let parser = char('A').right(char('B'));

let mut st = CharStream::new("ABC");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('B'), res);
assert_eq!("C", st.as_str());
assert_eq!(0, logs.len());

fn mid<B: 'f, C: 'f>(
self,
p1: impl Parsable<Stream = S, Result = B> + 'f,
p2: impl Parsable<Stream = S, Result = C> + 'f
) -> Parser<'f, B, S> where
Self: Sized + 'f,
S: Clone + 'f,

Combinator: mid

A sequential combinator that applys three parsers one after another. If all parsers succeed, returns the result of the second parser, otherwise fail.

Example

use naive_parsec::combinators::*;
use naive_parsec::core::Parsable;
use naive_parsec::primitives::{char, CharStream};

// Consume a character 'A', then a character 'B'
let parser = char('A').mid(char('B'), char('C'));

let mut st = CharStream::new("ABC");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('B'), res);
assert_eq!("", st.as_str());
assert_eq!(0, logs.len());

Loading content...